Project compiles

+++ siyana-renderer/src/OpenCLUtilities.cpp
@@ -598,7 +598,7 @@ void UpdateLocalPixels(void) {
 //	printf("OpenCL: Reading OpenCL device memory into local pixel array\n");
 	cl_int error = 0;
     const size_t origin[3] = {0, 0, 0};
-    const size_t region[3] = {width, height, 1};
+    const size_t region[3] = {static_cast<size_t>(width), static_cast<size_t>(height), 1};
         float* old_pixels = (float*)malloc(width * height * 4 * sizeof(float));
         memcpy(old_pixels, pixels, width * height * 4 * sizeof(float));
     //local pixel array should already be allocated
